Move model wrapper into its own module
